+++ mayan/apps/cabinets/search.py
@@ -51,31 +51,3 @@ search_model_cabinet.add_model_field(
 search_model_document.add_model_field(
     field='cabinets__label', label=_('Cabinets')
 )
-
-# Document file
-
-search_model_document_file.add_model_field(
-    field='document__cabinets__label',
-    label=_('Document cabinets')
-)
-
-# Document file page
-
-search_model_document_file_page.add_model_field(
-    field='document_file__document__cabinets__label',
-    label=_('Document cabinets')
-)
-
-# Document version
-
-search_model_document_version.add_model_field(
-    field='document__cabinets__label',
-    label=_('Document cabinets')
-)
-
-# Document version page
-
-search_model_document_version_page.add_model_field(
-    field='document_version__document__cabinets__label',
-    label=_('Document cabinets')
-)