+++ mayan/apps/document_parsing/search.py
@@ -5,6 +5,13 @@ from mayan.apps.documents.search import (
     search_model_document_file_page
 )
 
+# Document
+
+search_model_document.add_model_field(
+    field='files__file_pages__content__content',
+    label=_('Document file content')
+)
+
 # Document file
 
 search_model_document_file.add_model_field(
@@ -15,11 +22,4 @@ search_model_document_file.add_model_field(
 
 search_model_document_file_page.add_model_field(
     field='content__content', label=_('Document file content')
-)
-
-# Document
-
-search_model_document.add_model_field(
-    field='files__file_pages__content__content',
-    label=_('Document file content')
 )