+++ mayan/apps/file_metadata/search.py
@@ -26,14 +26,3 @@ search_model_document_file.add_model_field(
     field='file_metadata_drivers__entries__value',
     label=_('File metadata value')
 )
-
-# Document file page
-
-search_model_document_file_page.add_model_field(
-    field='document_file__file_metadata_drivers__entries__key',
-    label=_('File metadata key')
-)
-search_model_document_file_page.add_model_field(
-    field='document_file__file_metadata_drivers__entries__value',
-    label=_('File metadata value')
-)